+++ tests/test_taskgroups.py
@@ -30,9 +30,11 @@ from anyio import (
     current_effective_deadline,
     current_time,
     fail_after,
+    fail_at,
     get_cancelled_exc_class,
     get_current_task,
     move_on_after,
+    move_on_at,
     sleep,
     sleep_forever,
     wait_all_tasks_blocked,
@@ -714,6 +716,15 @@ async def test_fail_after_scope_cancelled_before_timeout() -> None:
         await checkpoint()
 
 
+async def test_fail_after_reason() -> None:
+    with pytest.raises(TimeoutError, match="oopsies"):
+        with fail_after(0, reason="oopsies") as scope:
+            await sleep(1)
+
+    assert scope.cancel_called
+    assert scope.cancelled_caught
+
+
 @pytest.mark.parametrize("delay", [0, 0.1], ids=["instant", "delayed"])
 async def test_move_on_after(delay: float) -> None:
     result = False
@@ -754,6 +765,63 @@ async def test_nested_move_on_after() -> None:
     assert outer_scope.cancelled_caught
     assert not inner_scope.cancel_called
     assert not inner_scope.cancelled_caught
+
+
+@pytest.mark.parametrize("delay", [0, 0.1], ids=["instant", "delayed"])
+async def test_fail_at(delay: float) -> None:
+    with pytest.raises(TimeoutError):
+        with fail_at(current_time() + delay) as scope:
+            try:
+                await sleep(1)
+            except get_cancelled_exc_class() as exc:
+                assert "deadline" in str(exc)
+                raise
+            else:
+                pytest.fail("sleep() should have raised a cancellation exception")
+
+    assert scope.cancel_called
+    assert scope.cancelled_caught
+
+
+async def test_fail_at_no_timeout() -> None:
+    with fail_at(None) as scope:
+        assert scope.deadline == float("inf")
+        await sleep(0.1)
+
+    assert not scope.cancel_called
+    assert not scope.cancelled_caught
+
+
+async def test_fail_at_reason() -> None:
+    with pytest.raises(TimeoutError, match="oopsies"):
+        with fail_at(current_time(), reason="oopsies") as scope:
+            await sleep(1)
+
+    assert scope.cancel_called
+    assert scope.cancelled_caught
+
+
+@pytest.mark.parametrize("delay", [0, 0.1], ids=["instant", "delayed"])
+async def test_move_on_at(delay: float) -> None:
+    result = False
+    with move_on_at(current_time() + delay) as scope:
+        await sleep(1)
+        result = True
+
+    assert not result
+    assert scope.cancel_called
+    assert scope.cancelled_caught
+
+
+async def test_move_on_at_no_timeout() -> None:
+    result = False
+    with move_on_at(None) as scope:
+        assert scope.deadline == float("inf")
+        await sleep(0.1)
+        result = True
+
+    assert result
+    assert not scope.cancel_called
 
 
 async def test_shielding() -> None:
@@ -2276,37 +2344,6 @@ class TestCreateTask:
         async with create_task_group() as tg:
             assert await tg.create_task(taskfunc(), context=ctx) == 42
 
-    async def test_task_name_default(self) -> None:
-        async def taskfunc() -> str | None:
-            return get_current_task().name
-
-        async with create_task_group() as tg:
-            handle = tg.create_task(taskfunc())
-            assert re.match(
-                r"<TaskHandle pending "
-                r"name='tests.test_taskgroups.TestCreateTask.test_task_name_default.<locals>.taskfunc' "
-                r"coro=<coroutine object(.+)>>",
-                repr(handle),
-            )
-            assert await handle == handle.name
-
-        assert (
-            handle.name
-            == "tests.test_taskgroups.TestCreateTask.test_task_name_default.<locals>.taskfunc"
-        )
-
-    async def test_task_name_custom_name(self) -> None:
-        async def taskfunc() -> None:
-            assert get_current_task().name == "custom name"
-
-        async with create_task_group() as tg:
-            handle = tg.create_task(taskfunc(), name="custom name")
-            assert handle.name == "custom name"
-            assert re.match(
-                r"<TaskHandle pending name='custom name' coro=<coroutine object(.+)>>",
-                repr(handle),
-            )
-
     async def test_wait(self) -> None:
         async def taskfunc() -> None:
             raise RuntimeError("dummy error")
@@ -2320,6 +2357,67 @@ class TestCreateTask:
         assert str(handle.exception) == "dummy error"
         assert handle.status is TaskHandle.Status.FAILED
         await handle.wait()
+
+
+@pytest.mark.parametrize("spawner", ["create_task", "start_soon", "start"])
+async def test_task_name_default(spawner: str) -> None:
+    async def taskfunc(
+        *, task_status: TaskStatus[None] = TASK_STATUS_IGNORED
+    ) -> str | None:
+        task_status.started()
+        return get_current_task().name
+
+    async with create_task_group() as tg:
+        match spawner:
+            case "create_task":
+                handle = tg.create_task(taskfunc())
+            case "start_soon":
+                handle = tg.start_soon(taskfunc)
+            case "start":
+                handle = await tg.start(taskfunc, return_handle=True)
+            case _:
+                raise ValueError
+
+        assert (
+            await handle
+            == handle.name
+            == "tests.test_taskgroups.test_task_name_default.<locals>.taskfunc"
+        )
+        assert re.match(
+            r"<TaskHandle finished "
+            r"name='tests.test_taskgroups.test_task_name_default.<locals>.taskfunc' "
+            r"coro=<coroutine object test_task_name_default.<locals>.taskfunc at 0x[0-9a-fA-F]+>>",
+            repr(handle),
+        )
+
+
+@pytest.mark.parametrize("spawner", ["create_task", "start_soon", "start"])
+async def test_task_name_custom_name(spawner: str) -> None:
+    async def taskfunc(
+        *, task_status: TaskStatus[None] = TASK_STATUS_IGNORED
+    ) -> str | None:
+        task_status.started()
+        return get_current_task().name
+
+    async with create_task_group() as tg:
+        match spawner:
+            case "create_task":
+                handle = tg.create_task(taskfunc(), name="custom name")
+            case "start_soon":
+                handle = tg.start_soon(taskfunc, name="custom name")
+            case "start":
+                handle = await tg.start(
+                    taskfunc, return_handle=True, name="custom name"
+                )
+            case _:
+                raise ValueError
+
+        assert await handle == handle.name == "custom name"
+        assert re.match(
+            r"<TaskHandle finished name='custom name' "
+            r"coro=<coroutine object test_task_name_custom_name.<locals>.taskfunc at 0x[0-9a-fA-F]+>>",
+            repr(handle),
+        )
 
 
 @pytest.mark.parametrize("create_task", [False, True])